compilation error on windows
get_rec_bits() must return uchar, it's result is often
assigned to uchar. And len is always less than 8.

+++ include/my_compare.h
@@ -91,7 +91,7 @@ typedef struct st_HA_KEYSEG		/* Key-portion */
 
 #define size_to_store_key_length(length) ((length) < 255 ? 1 : 3)
 
-static inline uint16 get_rec_bits(const uchar *ptr, uchar ofs, uint len)
+static inline uchar get_rec_bits(const uchar *ptr, uchar ofs, uint len)
 {
   uint16 val= ptr[0];
   if (ofs + len > 8)
